+++ src/main/java/name/divinityunbound/block/custom/WormholeTransporterBlock.java
@@ -38,6 +38,16 @@ public class WormholeTransporterBlock extends BlockWithEntity implements BlockEn
         return SHAPE;
     }
 
+    @Override
+    public boolean isShapeFullCube(BlockState state, BlockView world, BlockPos pos) {
+        return false;
+    }
+
+    @Override
+    public float getAmbientOcclusionLightLevel(BlockState state, BlockView world, BlockPos pos) {
+        return 1.0F;
+    }
+
     @Nullable
     @Override
     public BlockState getPlacementState(ItemPlacementContext ctx) {